Move has_error and get_queries
